surefire plugin updated, guice version updated, surefire configured to display stdout from native code, getVelocity methods reverted to int, shortest distance lib test package cleanup, ShortestPathsTestModule configured to load the map from test resources, instead using app map from config, logging of dists in Distance Matrix TP changed to trace
Former-commit-id: 2dd497688992d8e79dbcd1addeb24ed4e73b4967

+++ src/test/java/cz/agents/amodsim/ridesharing/vga/mock/TestOnDemandVehicle.java
@@ -47,7 +47,7 @@ public class TestOnDemandVehicle implements MovingEntity{
 	
 
 	@Override
-	public double getVelocity() {
+	public int getVelocity() {
 		return config.vehicleSpeedInMeters;
 	}
